bilateral_grid.py works as desired (JIt/compile_to_file and CPU/GPU variants)
Former-commit-id: c75b0a12d94af5f257c4d3f5a8ec80901ec1fc29

+++ boost_python_bindings/apps/blur.py
@@ -40,7 +40,7 @@ def get_input_data():
     rgb_data = imread(image_path)
     print("rgb_data", type(rgb_data), rgb_data.shape, rgb_data.dtype)
 
-    grey_data = np.mean(rgb_data, axis=2, dtype=float).astype(rgb_data.dtype)
+    grey_data = np.mean(rgb_data, axis=2, dtype=np.float32).astype(rgb_data.dtype)
     input_data = np.copy(grey_data, order="F")
 
     return input_data
